# docs/api/newsletter-openapi.yaml
openapi: 3.0.3
info:
  title: Newsletter API
  description: |
    Complete newsletter subscription management API for Next.js applications.
    
    Features:
    - Email subscription and unsubscription
    - Double opt-in confirmation
    - Analytics and statistics
    - Bulk operations
    - Multiple database backend support
    - Rate limiting and security
    
    Built for modern web applications with TypeScript support.
  version: 2.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/newsletter
    description: Production server
  - url: https://api-staging.example.com/newsletter
    description: Staging server
  - url: http://localhost:3000/api/newsletter
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {}
paths:
  /subscribe:
    post:
      tags:
        - Subscriptions
      summary: Subscribe to newsletter
      description: |
        Subscribe an email address to the newsletter. Supports additional metadata
        and tags for categorization. Implements rate limiting and duplicate checking.
      operationId: subscribe
      security:
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscribeRequest'
            examples:
              basic:
                summary: Basic subscription
                value:
                  email: user@example.com
              with_metadata:
                summary: With metadata and tags
                value:
                  email: user@example.com
                  source: landing_page
                  tags: ["updates", "releases"]
                  metadata:
                    referrer: homepage
                    campaign: launch_2024
      responses:
        '201':
          description: Successfully subscribed
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscribeResponse'
              examples:
                success:
                  summary: Successful subscription
                  value:
                    success: true
                    message: "Successfully subscribed!"
                    data:
                      subscription:
                        id: "123e4567-e89b-12d3-a456-426614174000"
                        email: "user@example.com"
                        status: "active"
                        source: "web"
                        tags: ["updates"]
                        subscribed_at: "2024-01-01T00:00:00.000Z"
                        confirmed: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already subscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                already_subscribed:
                  summary: Email already subscribed
                  value:
                    error: "Email already subscribed"
                    code: "ALREADY_SUBSCRIBED"
                    details:
                      subscribed_at: "2024-01-01T00:00:00.000Z"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for admin authentication
  headers:
    X-RateLimit-Limit:
      description: The number of allowed requests in the current period
      schema:
        type: integer
        example: 100
    X-RateLimit-Remaining:
      description: The number of remaining requests in the current period
      schema:
        type: integer
        example: 99
    X-RateLimit-Reset:
      description: The time at which the current rate limit window resets (Unix timestamp)
      schema:
        type: integer
        example: 1640995200
  schemas:
    SubscribeRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Valid email address to subscribe
          example: "user@example.com"
        source:
          type: string
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
          default: "web"
          description: Source of the subscription
          example: "landing_page"
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 30
            pattern: '^[a-zA-Z0-9_-]+$'
          description: Optional tags for categorization
          example: ["updates", "releases"]
        metadata:
          type: object
          description: Additional metadata (max 1KB when serialized)
          example:
            referrer: "homepage"
            campaign: "launch_2024"
            utm_source: "google"
    SubscribeResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Successfully subscribed!"
        data:
          type: object
          properties:
            subscription:
              $ref: '#/components/schemas/Subscription'
...
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "user@example.com"
        status:
          type: string
          enum: [pending, active, unsubscribed, bounced]
          example: "active"
        source:
          type: string
          example: "web"
        tags:
          type: array
          items:
            type: string
          example: ["updates", "releases"]
        metadata:
          type: object
          example:
            referrer: "homepage"
        subscribed_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
        confirmed_at:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-01T00:05:00.000Z"
        unsubscribed_at:
          type: string
          format: date-time
          nullable: true
          example: null
        confirmed:
          type: boolean
          example: true
        confirmation_token:
          type: string
          nullable: true
          example: "abc123xyz"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Validation failed"
        code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        details:
          type: object
          description: Additional error details
          example:
            field: "email"
            message: "Please enter a valid email address"
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2024-01-01T00:00:00.000Z"
        path:
          type: string
          description: Request path that caused the error
          example: "/api/newsletter/subscribe"
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Validation error
              value:
                error: "Validation failed"
                code: "VALIDATION_ERROR"
                details:
                  email: "Please enter a valid email address"
                timestamp: "2024-01-01T00:00:00.000Z"
                path: "/api/newsletter/subscribe"
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
            example: 300
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit:
              summary: Rate limit exceeded
              value:
                error: "Rate limit exceeded. Please wait before trying again."
                code: "RATE_LIMIT_EXCEEDED"
                details:
                  reset_time: 1640995200
                  retry_after: 300
                timestamp: "2024-01-01T00:00:00.000Z"
                path: "/api/newsletter/subscribe"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal_error:
              summary: Internal server error
              value:
                error: "Internal server error"
                code: "INTERNAL_ERROR"
                timestamp: "2024-01-01T00:00:00.000Z"
                path: "/api/newsletter/subscribe"
...